requirements.txt
ipywidgets
numba==0.58.1
papermill==2.6.0
dagit
dash==2.11.1
dtale==3.17.0
ydata-profiling==4.16.1
autoviz==0.1.905
sweetviz==2.3.1
yfinance==0.2.58
numpy<2.2
pandas<2.2
jupyter==1.1.1
scipy<1.11
psycopg2-binary==2.9.10